+++ generators/app/templates/statics/pages/models/global.js
@@ -1,0 +1,9 @@
+module.exports = () => {
+  return {
+    content: {
+      appTitle: '<%= appName %>',
+      pageTitle: '{content.appTitle}', // override this on a per route level
+      titleTag: '{content.appTitle} — {content.pageTitle}'
+    }
+  }
+}